[Feature] Support acr verification
This commit adds new options to oauth2-proxy to request and verify
various ACR values. This is intended as a replacement for the existing
--acr-values flag, while maintaining it's existence for backwards
compatibility.

The existing --acr-values flag submits the specified value as a
parameter of the login request, but does no verification of the
response. A relatively unsophisticated attacker could remove the
parameter from the url after being redirected by oauth2-proxy,
potentially bypassing additional verification, and oauth2-proxy would
complete the oauth exchange successfully.

When using the new arguments, the acr values as passed as a url
parameter and verified in the oauth response handler. Should the
response not provide a matching acr claim, the user is not authorized.